Reflect: when a property overshadows a method, dispatch to the
method in certain cases. For example, when the property is called
with 1+ arguments, or a block - a method call will be dispatched
instead of reading from the property.

This is possible because a property can't receive arguments, and
so - dispatching to the method being overshadowed seems to make the
most sense, and allows for methods like #method_missing to be shadowed
but remain functional when called with 2/3 args.